feat: router for better behavior, cleaner state handling, better modularity of components
