+++ components/sidebar/left.js
@@ -1,0 +1,7 @@
+export default function LeftSidebar() {
+  return (
+    <div className="h-full hidden md:block md:w-60 items-center justify-center bg-gray-100">
+      <h3> Left sidebar</h3>
+    </div>
+  );
+}
