+++ components/sidebar/right.js
@@ -1,0 +1,7 @@
+export default function RightSidebar() {
+  return (
+    <div className="h-full hidden md:block md:w-60 items-center justify-center bg-gray-100">
+      <h3> Right sidebar</h3>
+    </div>
+  );
+}
